+++ test/docs/debug.test.js
@@ -15,7 +15,11 @@ const Schema = mongoose.Schema;
  */
 
 const testSchema = new Schema({
-  dob: Date
+  dob: Date,
+  title: {
+    type: String,
+    required: false,
+  }
 }, {
   timestamps: {
     createdAt: 'created_at'
@@ -67,4 +71,19 @@ describe('debug: shell', function() {
     assert.equal(true, lastLog.includes('ISODate'));
   });
 
+  it('should allow to set the `debug` option on a per-connection basis (gh-12700)', async function() {
+    const m = new mongoose.Mongoose();
+    // `conn1` with active debug
+    const conn1 = m.createConnection(start.uri);
+    conn1.set('debug', true);
+    const testModel1 = conn1.model('Test', testSchema);
+    await testModel1.create({ dob: new Date(), title: 'Connection 1' });
+    const storedLog = lastLog;
+    // `conn2` without debug
+    const conn2 = m.createConnection(start.uri);
+    const testModel2 = conn2.model('Test', testSchema);
+    await testModel2.create({ dob: new Date(), title: 'Connection 2' });
+    // Last log should not have been overwritten
+    assert.equal(storedLog, lastLog);
+  });
 });