+++ test/document.test.js
@@ -9191,7 +9191,6 @@ describe('document', function() {
 
     const Test = db.model('Test', schema);
 
-
     const foo = new Test({ bar: 'bar' });
     await foo.save();
     assert.ok(!foo.isModified('bar'));
@@ -11016,6 +11015,54 @@ describe('document', function() {
     assert.equal(foo.get('bar.another'), 2);
   });
 
+  it('populating subdocument refs underneath maps throws (gh-12494) (gh-10856)', async function() {
+    // Bar model, has a name property and some other properties that we are interested in
+    const BarSchema = new Schema({
+      name: String,
+      more: String,
+      another: Number
+    });
+    const Bar = db.model('Bar', BarSchema);
+
+    // Denormalised Bar schema with just the name, for use on the Foo model
+    const BarNameSchema = new Schema({
+      _id: {
+        type: Schema.Types.ObjectId,
+        ref: 'Bar'
+      },
+      name: String
+    });
+
+    // Foo model, which contains denormalized bar data (just the name)
+    const FooSchema = new Schema({
+      something: String,
+      other: Number,
+      map: {
+        type: Map,
+        of: {
+          type: BarNameSchema,
+          ref: 'Bar'
+        }
+      }
+    });
+    const Foo = db.model('Foo', FooSchema);
+
+    const bar = await Bar.create({
+      name: 'I am Bar',
+      more: 'With more data',
+      another: 2
+    });
+    const { _id } = await Foo.create({
+      something: 'I am Foo',
+      other: 1,
+      map: { test: bar }
+    });
+
+    const err = await Foo.findById(_id).populate('map').then(() => null, err => err);
+    assert.ok(err);
+    assert.ok(err.message.includes('Cannot manually populate single nested subdoc underneath Map'), err.message);
+  });
+
   it('handles save with undefined nested doc under subdoc (gh-11110)', async function() {
     const testSchema = new Schema({
       level_1_array: [new Schema({
@@ -11927,6 +11974,106 @@ describe('document', function() {
     const rawDoc = await Test.collection.findOne({ _id });
     assert.ok(rawDoc);
     assert.deepStrictEqual(rawDoc.tags, ['mongodb']);
+  });
+
+  it('$clone() (gh-11849)', async function() {
+    const schema = new mongoose.Schema({
+      name: {
+        type: String,
+        validate: {
+          validator: (v) => v !== 'Invalid'
+        }
+      }
+    });
+    const Test = db.model('Test', schema);
+
+    const item = await Test.create({ name: 'Test' });
+
+    const doc = await Test.findById(item._id);
+    const clonedDoc = doc.$clone();
+
+    assert.deepEqual(clonedDoc, doc);
+    assert.deepEqual(clonedDoc._doc, doc._doc);
+    assert.deepEqual(clonedDoc.$__, doc.$__);
+
+    // Editing a field in the cloned doc does not effect
+    // the original doc
+    clonedDoc.name = 'Test 2';
+    assert.equal(doc.name, 'Test');
+    assert.equal(clonedDoc.name, 'Test 2');
+    assert.ok(!doc.$isModified('name'));
+    assert.ok(clonedDoc.$isModified('name'));
+
+    // Saving the cloned doc does not effect `modifiedPaths`
+    // in the original doc
+    const modifiedPaths = [...doc.modifiedPaths()];
+    await clonedDoc.save();
+    assert.deepEqual(doc.modifiedPaths(), modifiedPaths);
+
+    // Cloning a doc with invalid field preserve the
+    // invalid field value
+    doc.name = 'Invalid';
+    await assert.rejects(async() => {
+      await doc.validate();
+    });
+
+    await clonedDoc.validate();
+
+    const invalidClonedDoc = doc.$clone();
+    doc.name = 'Test';
+    await doc.validate();
+    await assert.rejects(async() => {
+      await invalidClonedDoc.validate();
+    });
+
+    // Setting a session on the cloned doc does not
+    // affect the session in the original doc
+    const session = await Test.startSession();
+    clonedDoc.$session(session);
+    assert.strictEqual(doc.$session(), null);
+    assert.strictEqual(clonedDoc.$session(), session);
+  });
+
+  it('can create document with document array and top-level key named `schema` (gh-12480)', async function() {
+    const AuthorSchema = new Schema({
+      fullName: { type: 'String', required: true }
+    });
+
+    const BookSchema = new Schema({
+      schema: { type: 'String', required: true },
+      title: { type: 'String', required: true },
+      authors: [AuthorSchema]
+    }, { supressReservedKeysWarning: true });
+
+    const Book = db.model('Book', BookSchema);
+
+    await Book.create({
+      schema: 'design',
+      authors: [{ fullName: 'Sourabh Bagrecha' }],
+      title: 'The power of JavaScript'
+    });
+  });
+
+  it('handles setting array to itself after saving and pushing a new value (gh-12656)', async function() {
+    const Test = db.model('Test', new Schema({
+      list: [{
+        a: Number
+      }]
+    }));
+    await Test.create({ list: [{ a: 1, b: 11 }] });
+
+    let doc = await Test.findOne();
+    doc.list.push({ a: 2 });
+    doc.list = [...doc.list];
+    await doc.save();
+
+    doc.list.push({ a: 3 });
+    doc.list = [...doc.list];
+    await doc.save();
+
+    doc = await Test.findOne();
+    assert.equal(doc.list.length, 3);
+    assert.deepStrictEqual(doc.list.map(el => el.a), [1, 2, 3]);
   });
 });