+++ test/model.populate.test.js
@@ -10802,5 +10802,89 @@ describe('model: populate:', function() {
       assert.ok(err);
       assert.ok(err.message.includes('strictPopulate'), err.message);
     });
+
+    it('allows overwriting localField and foreignField when populating a virtual gh-6963', async function() {
+      const testSchema = Schema({ name: String, uuid: 'ObjectId' }, { toJSON: { virtuals: true }, toObject: { virtuals: true } });
+      const userSchema = Schema({ name: String, field: { type: mongoose.Schema.Types.ObjectId, ref: 'gh6963' }, change: { type: mongoose.Schema.Types.ObjectId, ref: 'gh6963' } });
+      testSchema.virtual('test', {
+        ref: 'gh6963-2',
+        localField: '_id',
+        foreignField: 'field'
+      });
+
+      const Test = db.model('gh6963', testSchema);
+      const User = db.model('gh6963-2', userSchema);
+
+      const entry = await Test.create({
+        name: 'Test',
+        uuid: mongoose.Types.ObjectId()
+      });
+      const otherEntry = await Test.create({
+        name: 'Other Test',
+        uuid: mongoose.Types.ObjectId()
+      });
+      await User.create({
+        name: 'User',
+        field: entry._id,
+        change: otherEntry._id
+      });
+      const res = await Test.findOne({ _id: otherEntry._id }).populate({ path: 'test', foreignField: 'change' });
+      const other = await Test.findOne({ _id: entry._id }).populate({ path: 'test', foreignField: 'change' });
+      assert.equal(res.test.length, 1);
+      assert.equal(other.test.length, 0);
+      // make sure its not broken
+      const response = await Test.findOne({ _id: entry._id }).populate('test');
+      assert.equal(response.test.length, 1);
+      // =================localField======================
+      const localEntry = await Test.create({
+        name: 'local test',
+        uuid: mongoose.Types.ObjectId()
+      });
+      const otherLocalEntry = await Test.create({
+        name: 'other local test',
+        uuid: mongoose.Types.ObjectId()
+      });
+
+      await User.create({
+        name: 'local user',
+        field: localEntry.uuid,
+        change: otherLocalEntry.uuid
+      });
+
+      const localTest = await Test.
+        findOne({ _id: localEntry._id }).
+        populate({ path: 'test', localField: 'uuid' });
+      assert.equal(localTest.test.length, 1);
+      const otherLocalTest = await Test.
+        findOne({ _id: otherLocalEntry._id }).
+        populate({ path: 'test', localField: 'uuid' });
+      assert.equal(otherLocalTest.test.length, 0);
+      // should be empty because the original local field was _id and we created a doc with uuids
+      const check = await Test.
+        findOne({ _id: localEntry._id }).
+        populate({ path: 'test' });
+      assert.equal(check.test.length, 0);
+      // ============localFieldAndForeignField============
+      const bothEntry = await Test.create({ name: 'Both', uuid: mongoose.Types.ObjectId() });
+      const otherBothEntry = await Test.create({ name: 'Other Both', uuid: mongoose.Types.ObjectId() });
+      await User.create({
+        name: 'both user',
+        field: bothEntry.uuid,
+        change: otherBothEntry.uuid
+      });
+      const bothTest = await Test.
+        findOne({ _id: otherBothEntry._id }).
+        populate({ path: 'test', localField: 'uuid', foreignField: 'change' });
+      assert.equal(bothTest.test.length, 1);
+      const otherBothTest = await Test.
+        findOne({ _id: bothEntry._id }).
+        populate({ path: 'test', localField: 'uuid', foreignField: 'change' });
+      assert.equal(otherBothTest.test.length, 0);
+      const normal = await Test.
+        findOne({ _id: otherBothEntry._id }).
+        populate({ path: 'test' });
+      // should be empty because the original local field was _id and we created a doc with uuids
+      assert.equal(normal.test.length, 0);
+    });
   });
 });