+++ test/types.map.test.js
@@ -1050,4 +1050,56 @@ describe('Map', function() {
     const res = doc.toObject({ flattenMaps: true });
     assert.equal(res.l1.l1key.l2.l2key.value, 'abc');
   });
+
+  it('handles populating map of arrays (gh-12494)', async function() {
+    const User = new mongoose.Schema({
+      name: String,
+      addresses: {
+        type: Map,
+        of: [{
+          type: mongoose.Schema.Types.ObjectId,
+          ref: 'Address'
+        }],
+        default: {}
+      }
+    });
+
+    const Address = new mongoose.Schema({
+      city: String
+    });
+
+    const UserModel = db.model('User', User);
+    const AddressModel = db.model('Address', Address);
+
+    const address = await AddressModel.create({ city: 'London' });
+
+    const { _id } = await UserModel.create({
+      name: 'Name',
+      addresses: {
+        home: [address._id]
+      }
+    });
+
+    // Using `.$*`
+    let query = UserModel.findById(_id);
+    query.populate({
+      path: 'addresses.$*'
+    });
+
+    let doc = await query.exec();
+    assert.ok(Array.isArray(doc.addresses.get('home')));
+    assert.equal(doc.addresses.get('home').length, 1);
+    assert.equal(doc.addresses.get('home')[0].city, 'London');
+
+    // Populating just one path in the map
+    query = UserModel.findById(_id);
+    query.populate({
+      path: 'addresses.home'
+    });
+
+    doc = await query.exec();
+    assert.ok(Array.isArray(doc.addresses.get('home')));
+    assert.equal(doc.addresses.get('home').length, 1);
+    assert.equal(doc.addresses.get('home')[0].city, 'London');
+  });
 });